+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - add continuous sound effects for c-dash and dive
+- add animated conveyor belts
 
 ## [0.0.15] - 2024-02-28